refactor(System): RHICOMPL-1139 system profiles to concern
